Deterministic Order of Nodes when Serializing Objects
This uses a feature for ordering Properties during Serialization that is not part of the official FileDBSerializer as of the time of this commit.

Also, contrary to the fixed export order, the validity checker in the test program now uses XMLDiff to ignore different Node order since some of the original FileDB files from the game have variations in their node order, which get removed during a round trip test.